Mass testing
Bumped timeout to 30s so the mass testing on the index page worked.

Uncommented all the other news feeds on the index page.

Removed Working Adults feed from index page because it won't be used in the element since the site is now MercerForYou.com and that's not in CommonSpot.